Move eth_getProof to eth rpc
