+++ crates/ethereum-rpc/Cargo.toml
@@ -38,6 +38,7 @@ reth-rpc-types-compat = { workspace = true }
 sov-db = { path = "../../crates/sovereign-sdk/full-node/db/sov-db" }
 sov-modules-api = { path = "../sovereign-sdk/module-system/sov-modules-api", default-features = false }
 sov-rollup-interface = { path = "../sovereign-sdk/rollup-interface", features = ["native"] }
+sov-state = { path = "../sovereign-sdk/module-system/sov-state", default-features = false }
 
 [dev-dependencies]
 proptest = { workspace = true }
